New delegates and private helper methods for ActionEvaluator
New delegates

Renaming

Renaming

Private helper methods introduced

Helper methods parameters changed

Merge conflict resolve

Merge conflict resolve

Helper method usage propagation

GetGetter methods combined
