+++ Libplanet/Action/AccountStateGetter.cs
@@ -1,5 +1,7 @@
 #nullable enable
 using Bencodex.Types;
+using Libplanet.Blockchain;
+using Libplanet.Blocks;
 
 namespace Libplanet.Action
 {
@@ -16,4 +18,10 @@ namespace Libplanet.Action
     /// <returns>The account state if exists.  Otherwise <c>null</c>.
     /// </returns>
     public delegate IValue? AccountStateGetter(Address address);
+
+    public delegate IValue? StateGetter<T>(
+        Address address,
+        BlockHash? hashDigest,
+        StateCompleter<T> stateCompleter)
+            where T : IAction, new();
 }
